refactor : personality 타입 분리
