+++ app/join/funnels/step5/components/PersonalitySection.client.tsx
@@ -1,6 +1,6 @@
 import PersonalityItemServer from './PersonalityItem.server';
-import { PersonalityType } from '../type';
 import { personalityList } from '@/constants/personalityList';
+import { PersonalityType } from '@/types/profile';
 
 export default function PersonalitySection() {
   return (
