+++ types/profile.ts
@@ -1,0 +1,5 @@
+export type PersonalityType = {
+  id: number;
+  keyword: string;
+  keywordInEnglish: string;
+};
